+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+* `version` module with
+  * `version.get_versions()` returning an OrderedDict of library:version inclusing libzim
+  * `version.print_versions()` print it on stdout (or another fd)
+  * `version.get_libzim_version()` returns the libzim version only
+
 ## [2.1.0] - 2022-12-06
 
 ### Added